requirements.txt
requests
matplotlib
tk
binanc